# src/backend/core/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "messages API",
        "version": "1.0.0 (v1.0)",
        "description": "This is the messages API schema."
    },
    "paths": {
...
        "/api/v1.0/labels/": {
            "get": {
                "operationId": "labels_list",
                "description": "\n        List all labels accessible to the user in a hierarchical structure.\n\n        The response returns labels in a tree structure where:\n        - Labels are ordered alphabetically by name\n        - Each label includes its children (sub-labels)\n        - The hierarchy is determined by the label's name (e.g., \"Inbox/Important\" is a child of \"Inbox\")\n\n        You can filter labels by mailbox using the mailbox_id query parameter.\n        ",
                "parameters": [
                    {
                        "in": "query",
                        "name": "mailbox_id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "\n                Filter labels by mailbox ID. If not provided, returns labels from all accessible mailboxes.\n                "
                    }
                ],
                "tags": [
                    "labels"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TreeLabel"
                                    }
                                }
                            }
                        },
                        "description": "List of labels in hierarchical structure"
                    }
                }
            },
            "post": {
                "operationId": "labels_create",
                "description": "View and manage labels",
                "tags": [
                    "labels"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LabelRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/LabelRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TreeLabel"
                                    }
                                }
                            }
                        },
                        "description": "Created labels in hierarchical structure"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "Validation error"
                                }
                            }
                        },
                        "description": "Invalid input data"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "You need EDITOR or ADMIN role to manage labels"
                                }
                            }
                        },
                        "description": "Permission denied"
                    }
                }
            }
        },
        "/api/v1.0/labels/{id}/": {
            "put": {
                "operationId": "labels_update",
                "description": "View and manage labels",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "labels"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LabelRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/LabelRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Label"
                                }
                            }
                        },
                        "description": "Label updated successfully"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "Validation error"
                                }
                            }
                        },
                        "description": "Invalid input data"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "You need EDITOR or ADMIN role to manage labels"
                                }
                            }
                        },
                        "description": "Permission denied"
                    }
                }
            },
            "patch": {
                "operationId": "labels_partial_update",
                "description": "View and manage labels",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "labels"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedLabelRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedLabelRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Label"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "labels_destroy",
                "description": "View and manage labels",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "labels"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Label deleted successfully"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "You need EDITOR or ADMIN role to manage labels"
                                }
                            }
                        },
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Label not found"
                    }
                }
            }
        },
        "/api/v1.0/labels/{id}/add-threads/": {
            "post": {
                "operationId": "labels_add_threads_create",
                "description": "View and manage labels",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "labels"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "thread_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "description": "List of thread IDs to add to this label"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Label"
                                }
                            }
                        },
                        "description": "Threads added to label successfully"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "Validation error"
                                }
                            }
                        },
                        "description": "Invalid input data"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "You need EDITOR or ADMIN role to manage labels"
                                }
                            }
                        },
                        "description": "Permission denied"
                    }
                }
            }
        },
        "/api/v1.0/labels/{id}/remove-threads/": {
            "post": {
                "operationId": "labels_remove_threads_create",
                "description": "View and manage labels",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "labels"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "thread_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "description": "List of thread IDs to remove from this label"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Label"
                                }
                            }
                        },
                        "description": "Threads removed from label successfully"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "Validation error"
                                }
                            }
                        },
                        "description": "Invalid input data"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "detail": "You need EDITOR or ADMIN role to manage labels"
                                }
                            }
                        },
                        "description": "Permission denied"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LabelRequest": {
                "type": "object",
                "description": "Serializer for Label model.",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Name of the label/folder (can use slashes for hierarchy, e.g. 'Work/Projects')",
                        "maxLength": 255
                    },
                    "color": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Color of the label in hex format (e.g. #FF0000)",
                        "maxLength": 7
                    },
                    "mailbox": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Mailbox that owns this label"
                    },
                    "threads": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Threads that have this label"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the label, used by AI to understand its purpose",
                        "maxLength": 255
                    },
                    "auto_labellisation": {
                        "type": "boolean",
                        "description": "Whether this label should be automatically applied by AI"
                    }
                },
                "required": [
                    "mailbox",
                    "name"
                ]
            },
...
            "TreeLabel": {
                "type": "object",
                "description": "Serializer for tree label response structure (OpenAPI purpose only...).",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "slug": {
                        "type": "string",
                        "readOnly": true
                    },
                    "color": {
                        "type": "string",
                        "readOnly": true
                    },
                    "display_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "description": {
                        "type": "string",
                        "readOnly": true
                    },
                    "auto_labellisation": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TreeLabel"
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "auto_labellisation",
                    "children",
                    "color",
                    "description",
                    "display_name",
                    "id",
                    "name",
                    "slug"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}